% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R, R/filter.R
\name{classify_tile}
\alias{classify_tile}
\alias{filter_annual_samples}
\title{Classify a Data Cube tile}
\usage{
classify_tile(cube, model, ..., smooth_type = "bayes", roi = NULL)

filter_annual_samples(samples_ts, n_instances = 24)
}
\arguments{
\item{cube}{a \code{sits_cube} object with the cube metadata.}

\item{model}{a \code{sits_model} object with the machine learning
model.}

\item{...}{additional parameters.}

\item{smooth_type}{a \code{character} with the type of smoothing.}

\item{roi}{Region of interest. It can be an \code{sf_object}, a
\code{shapefile} , or a bounding box vector with
named XY values ("xmin", "xmax", "ymin", "ymax") or
named lat/long values
("lon_min", "lat_min", "lon_max", "lat_max").}

\item{samples_ts}{a \code{sits_tibble} with the time series with the
training samples.}

\item{n_instances}{a \code{numeric} with the number of instances to be
filtered.}
}
\value{
a \code{list} with the probabilities, smoothed and labelled cubes.

a \code{sits_tibble} with the time series filtered.
}
\description{
This function produces the complete cycle of a classification
of a cube. It generates the probability to the labels of a map.

This function filters each time series by the predetermined
number of instances.
}
