% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{sample_temporal_extents}
\alias{sample_temporal_extents}
\title{Sample temporal extents}
\usage{
sample_temporal_extents(samples)
}
\arguments{
\item{samples}{a \code{sits_tibble} with the time series with the
training samples.}
}
\value{
a \code{data.frame} with start_date and end_date of each time series
period.
}
\description{
This function creates a data.frame with the start and end dates
of each time series period.
}
